Extract Unhookd::Notifiers::Slack object.
